Rework object allocation to avoid undefined behavior
This makes it a bit more complicated to work out the totalObjectSize, because it has to be passed into the VMObject constructor, but it also works out some other odities.

- make additional bytes arg to `new` operator mandatory, just for consistency
- change naming of various bits
- use constructors more aggressively to initialize VM* fields and avoid unnecessary setters
- change VMArray to use numberOfFields to indexed elements. This means we do not support any fields on subclasses of array, which is fine with me. We don’t really support subclassing array in most SOMs anyway. And this way, VMArray is really just a VMObject from the perspective of GC

+++ src/vm/Universe.h
@@ -64,7 +64,7 @@ public:
     void Assert(bool) const;
 
     //VMObject instanciation methods. These should probably be refactored to a new class
-    VMArray* NewArray(long) const;
+    VMArray* NewArray(size_t) const;
 
     VMArray* NewArrayList(std::vector<vm_oop_t>& list) const;
     VMArray* NewArrayList(std::vector<VMInvokable*>& list) const;
